tests: fix type of free_recvbuffs()
Fixes a 32-bit RasPi warning.

+++ tests/libntp/recvbuff.c
@@ -23,7 +23,7 @@ TEST(recvbuff, Initialization) {
 }
 
 TEST(recvbuff, GetAndFree) {
-	int initial = free_recvbuffs();
+	u_long initial = free_recvbuffs();
 	recvbuf_t* buf = get_free_recv_buffer();
 
 	TEST_ASSERT_EQUAL(initial-1, free_recvbuffs());
